Address embedding-provider code review
- RHEEDEmbeddingResult stores cluster_sizes as float64 so partial/missing
  cluster_size coerces to NaN instead of crashing np.asarray (int64 has no
  missing-value sentinel); + regression test.
- to_result requires point["index"] (raise on malformed data rather than
  silently indexing rows with None).
- _expand_low_level_features pins json_normalize(max_level=1) so a future
  nested feature value can't silently dot-explode into extra columns.

+++ src/atomscale/results/rheed_embedding.py
@@ -65,8 +65,10 @@ class RHEEDEmbeddingResult(MSONable):
         self.unix_time_ms = (
             np.asarray(unix_time_ms, dtype=np.float64) if unix_time_ms is not None else None
         )
+        # float64 (not int64) so a partial/missing cluster_size (None) coerces to
+        # NaN rather than raising at array construction.
         self.cluster_sizes = (
-            np.asarray(cluster_sizes, dtype=np.int64) if cluster_sizes is not None else None
+            np.asarray(cluster_sizes, dtype=np.float64) if cluster_sizes is not None else None
         )
         self.count = count if count is not None else int(self.vectors.shape[0])
         self.truncated = truncated
